# services/adapter-fabric/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TANTU Adapter Fabric",
    "description": "Multi-protocol shopfloor ingestion \u2014 OPC-UA, Modbus, MQTT, MTConnect, EtherNet/IP, Camera \u2014 normalized to one schema, derived DefectEvents only (no image field).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Issue JWT (demo)",
        "operationId": "auth_token_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenIn": {
        "properties": {
          "sub": {
            "type": "string",
            "title": "Sub",
            "default": "operator"
          },
          "plant_id": {
            "type": "string",
            "title": "Plant Id",
            "default": "plant-demo-01"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "engineer"
          },
          "exp_min": {
            "type": "integer",
            "title": "Exp Min",
            "default": 60
          }
        },
        "type": "object",
        "title": "TokenIn"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}